pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.2.0",
    "flask-uploads>=0.2.1",
    "flask>=3.1.0",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "numpy>=2.2.4",
    "psycopg2-binary>=2.9.10",
    "sqlalchemy>=2.0.39",
    "werkzeug>=3.1.3",
    "pillow>=11.1.0",
    "scikit-image>=0.25.2",
    "scikit-learn>=1.6.1",
    "imageio>=2.37.0",
    "flask-reuploaded>=1.4.0",
]
